REF make rich an optional dependency (#47)
