+++ mlresearch/_min_dependencies.py
@@ -14,8 +14,8 @@ dependent_packages = {
     "numpy": (NUMPY_MIN_VERSION, "install"),
     "scikit-learn": (SKLEARN_MIN_VERSION, "install"),
     "imbalanced-learn": (IMBLEARN_MIN_VERSION, "install"),
-    "rich": (RICH_MIN_VERSION, "install"),
     "requests": ("2.26.0", "install"),
+    "rich": (RICH_MIN_VERSION, "optional"),
     "matplotlib": (MATPLOTLIB_MIN_VERSION, "optional, tests"),
     "research-learn": ("0.3.0", "tests"),
     "pytest-cov": ("3.0.0", "tests"),
